requirements.txt
streamlit
opencv-python
pillow
imagehash
beautifulsoup4
requests
lxml